backend/requirements.txt
# Agrawal Estate Planner - Backend Dependencies

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Data Validation
pydantic==2.10.2
pydantic-settings==2.6.1

# File Processing
pandas==2.2.3
openpyxl==3.1.5          # Excel file support
pdfplumber==0.11.4       # PDF parsing for statements
python-multipart==0.0.17  # File uploads

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
python-dateutil==2.9.0
httpx==0.28.0            # HTTP client for testing
requests==2.32.3         # HTTP client for notifications
apscheduler==3.10.4      # Advanced Python Scheduler for background tasks
pytz==2024.2             # Timezone support
yfinance==0.2.47         # Yahoo Finance API for stock/options data

# Notifications (optional)
twilio==9.3.5            # For WhatsApp notifications via Twilio (optional)

# AI/Vision for screenshot parsing
openai==1.55.0           # OpenAI API for vision/OCR

# Development
pytest==8.3.4
pytest-asyncio==0.24.0

